+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.3.0 (Next)
 - Support Vault Kubernetes authentication method.
 - `auth_engine` source parameter now required and not optional.
+- Generalize authentication parameters to `auth_mount` and `vault_role`.
 
 ### 1.2.4
 - Minor code optimization.